Bug 502090 - [Cocoa] Bridge support files should be updated to latest macOS supported version
Source changes generated by the MacGenerator using the updated bridge
support files from macOS 10.10

+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/NSOpenGLContext.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2012 IBM Corporation and others.
+ * Copyright (c) 2000, 2017 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -50,7 +50,7 @@ public void makeCurrentContext() {
 	OS.objc_msgSend(this.id, OS.sel_makeCurrentContext);
 }
 
-public void setValues(int[] vals, int param) {
+public void setValues(int[] vals, long /*int*/ param) {
 	OS.objc_msgSend(this.id, OS.sel_setValues_forParameter_, vals, param);
 }
